update to latest git: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@40522 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/nmap/nsock/src/nsock_internal.h
@@ -5,7 +5,7 @@
  *                                                                         *
  ***********************IMPORTANT NSOCK LICENSE TERMS***********************
  *                                                                         *
- * The nsock parallel socket event library is (C) 1999-2018 Insecure.Com   *
+ * The nsock parallel socket event library is (C) 1999-2019 Insecure.Com   *
  * LLC This library is free software; you may redistribute and/or          *
  * modify it under the terms of the GNU General Public License as          *
  * published by the Free Software Foundation; Version 2.  This guarantees  *
@@ -54,7 +54,7 @@
  *                                                                         *
  ***************************************************************************/
 
-/* $Id: nsock_internal.h 37126 2018-01-28 21:18:17Z fyodor $ */
+/* $Id$ */
 
 #ifndef NSOCK_INTERNAL_H
 #define NSOCK_INTERNAL_H
@@ -473,7 +473,7 @@ void event_delete(struct npool *nsp, struct nevent *nse);
  * etc. */
 void nsock_pool_add_event(struct npool *nsp, struct nevent *nse);
 
-void nsock_connect_internal(struct npool *ms, struct nevent *nse, int type, int proto, struct sockaddr_storage *ss, size_t sslen, unsigned short port);
+void nsock_connect_internal(struct npool *ms, struct nevent *nse, int type, int proto, struct sockaddr_storage *ss, size_t sslen, unsigned int port);
 
 /* Comments on using the following handle_*_result functions are available in nsock_core.c */
